pyproject.toml
[tool.poetry]
name = "lido-trustless-tlv-solution"
version = "0.1.0"
description = ""
authors = []
license = "GPL 3.0"
readme = ["README.md"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
# conflicts with ssz, but is installed separately via `pipx install eth-brownie` so just documenting usage/dependency
#eth-brownie = "^1.19.3"
ssz = "^0.3.1"
flask = "^2.3.2"
typed-argparse = "^0.2.11"
jinja2 = "^3.1.2"
requests = "^2.31.0"
validators = "^0.22.0"
python-dotenv = "^1.0.0"
psutil = "^5.9.5"

[tool.poetry.group.dev.dependencies]
pylint = "^2.0.0"
hypothesis = "^6.0.0"
black = "^22.0.0"  # brownie needs <23

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "unit: tests with using mocks and don't make external requests",
    "integration: tests with using providers",
    "possible_integration: tests with using providers, but can be run using mocks",
    "e2e: complex tests with using providers and real Ethereum network",
]
addopts = "-s --pdbcls pudb.debugger:Debugger"

[tool.coverage.run]
branch = true

[tool.pylint.format]
max-line-length = "120"

[tool.pylint."messages control"]
disable = [
    # Disabled by default
    "raw-checker-failed",
    "bad-inline-option",
    "locally-disabled",
    "file-ignored",
    "suppressed-message",
    "useless-suppression",
    "deprecated-pragma",
    "use-symbolic-message-instead",
    # Disabled by us
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "too-few-public-methods",
    "too-many-instance-attributes",
    "too-many-arguments",
    "too-many-locals",
    "invalid-envvar-default",
]

[tool.mypy]
strict_equality = true
mypy_path = "stubs"

[tool.black]
line-length = 120
skip-string-normalization = true
